requirements.txt
Flask
faiss-cpu
numpy
requests
sentence-transformers
gunicorn
